feat(web): lift session header so label stays editable after voting
The item label and round eyebrow previously lived inside Vote.jsx and
Results.jsx respectively, which meant the host's label input unmounted
the moment any vote was cast and remounted only after the next reset.
A host who realised the label needed correcting between reveal and
"Next Item" had no way to fix it.

This change extracts both into a new SessionHeader container that
GamePane mounts above the conditional Vote/Results region. The label
TextField (host) and italic display (non-host) plus the "Round N · Revealed"
eyebrow now persist across the voted=false / voted=true transition,
so the host can edit the label at any point in the round.

The eyebrow's round number is also now derived as `rounds.length + 1`
(the round currently being played) rather than the old "completed +
in-flight" calculation, which read as "Results" on a fresh session.

Tests:
- New SessionHeader.test.jsx covers eyebrow rendering, host vs non-host
  label UI, debounced /setLabel POST, and the new "host can edit label
  after voting" regression.
- Vote.test.jsx and Results.test.jsx drop the now-orphan label and
  eyebrow tests respectively.
- session-labels-csv.spec.js's "label displays on results screen after
  voting" updates the host assertion from getByText to toHaveValue,
  reflecting that the host's label is now in a persistent input.
