+++ planningpoker-web/src/containers/Vote.jsx
@@ -1,10 +1,8 @@
-import { useState, useEffect, useRef } from 'react'
+import { useState } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
-import TextField from '@mui/material/TextField'
-import InputAdornment from '@mui/material/InputAdornment'
-import { vote, voteOptimistic, setLabel } from '../actions'
+import { vote, voteOptimistic } from '../actions'
 import UsersTable from './UsersTable'
 
 function cardSx(isSelected, isDisabled) {
@@ -51,65 +49,8 @@ export default function Vote() {
   const dispatch = useDispatch()
   const sessionId = useSelector((state) => state.game.sessionId)
   const playerName = useSelector((state) => state.game.playerName)
-  const isAdmin = useSelector((state) => state.game.isAdmin)
-  const currentLabel = useSelector((state) => state.game.currentLabel)
   const legalEstimates = useSelector((state) => state.game.legalEstimates)
   const [selected, setSelected] = useState(null)
-  const [labelInput, setLabelInput] = useState(currentLabel)
-  const [lastBroadcastLabel, setLastBroadcastLabel] = useState(currentLabel)
-  const [justSaved, setJustSaved] = useState(false)
-  const debounceTimerRef = useRef(null)
-  const savedTimerRef = useRef(null)
-
-  useEffect(() => {
-    setLabelInput(currentLabel)
-    setLastBroadcastLabel(currentLabel)
-  }, [currentLabel])
-
-  useEffect(() => {
-    return () => {
-      if (debounceTimerRef.current) clearTimeout(debounceTimerRef.current)
-      if (savedTimerRef.current) clearTimeout(savedTimerRef.current)
-    }
-  }, [])
-
-  const commitLabel = (nextValue) => {
-    const value = nextValue !== undefined ? nextValue : labelInput
-    if (value === lastBroadcastLabel) return
-    dispatch(setLabel(playerName, sessionId, value))
-    setLastBroadcastLabel(value)
-    setJustSaved(true)
-    if (savedTimerRef.current) clearTimeout(savedTimerRef.current)
-    savedTimerRef.current = setTimeout(() => setJustSaved(false), 1500)
-  }
-
-  const handleLabelChange = (e) => {
-    const next = e.target.value
-    setLabelInput(next)
-    if (debounceTimerRef.current) clearTimeout(debounceTimerRef.current)
-    debounceTimerRef.current = setTimeout(() => {
-      commitLabel(next)
-    }, 1000)
-  }
-
-  const handleBlur = () => {
-    if (debounceTimerRef.current) {
-      clearTimeout(debounceTimerRef.current)
-      debounceTimerRef.current = null
-    }
-    commitLabel()
-  }
-
-  const handleKeyDown = (e) => {
-    if (e.key === 'Enter') {
-      e.preventDefault()
-      if (debounceTimerRef.current) {
-        clearTimeout(debounceTimerRef.current)
-        debounceTimerRef.current = null
-      }
-      commitLabel()
-    }
-  }
 
   const doVote = (val) => {
     if (selected) return
@@ -137,41 +78,6 @@ export default function Vote() {
         }}
       >
         <Box>
-          <Box sx={{ display: 'flex', alignItems: 'center', gap: 1, mb: 2, minHeight: 40 }}>
-            <Box sx={{ flex: 1 }}>
-              {isAdmin ? (
-                <TextField
-                  variant="standard"
-                  placeholder="Item label (optional)"
-                  fullWidth
-                  size="small"
-                  value={labelInput}
-                  onChange={handleLabelChange}
-                  onBlur={handleBlur}
-                  onKeyDown={handleKeyDown}
-                  inputProps={{ maxLength: 100 }}
-                  InputProps={{
-                    endAdornment: justSaved ? (
-                      <InputAdornment position="end">
-                        <Typography
-                          variant="caption"
-                          sx={{ color: 'success.main', whiteSpace: 'nowrap' }}
-                        >
-                          ✓ Saved
-                        </Typography>
-                      </InputAdornment>
-                    ) : null,
-                  }}
-                />
-              ) : (
-                currentLabel && (
-                  <Typography variant="body2" color="text.secondary" sx={{ fontStyle: 'italic' }}>
-                    {currentLabel}
-                  </Typography>
-                )
-              )}
-            </Box>
-          </Box>
           <Box
             sx={{
               display: 'grid',
